# swagger_output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "SGS Bacen API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://localhost:3001/"
    }
  ],
  "paths": {
    "/bacen/seriesChart/{code}": {
      "post": {
        "summary": "Cria no console um gráfico de uma série do BACEN",
        "description": "Cria no console um gráfico de uma série do BACEN",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Código de série temporal do BACEN",
            "example": 226
          }
        ],
        "responses": {
          "200": {
            "description": "Gráfico desenhado! Verifique o console do NodeJS."
          },
          "400": {
            "description": "Deu ruim."
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangedDate"
              }
            }
          }
        }
      }
    },
    "/bacen/metric/{code}": {
      "get": {
        "summary": "Traz o valor de uma série do BACEN",
        "description": "Dado um código válido, retorna o valor de uma série do BACEN para a data requerida.",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Código de série temporal do BACEN",
            "example": 226
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RangedDate"
              }
            }
          }
        }
      }
    },
    "/bacen/metric/info/{code}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "Metric": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "226"
          }
        }
      },
      "SingleDate": {
        "type": "string",
        "example": "15/08/2022"
      },
      "RangedDate": {
        "type": "object",
        "properties": {
          "startDate": {
            "type": "string",
            "example": "01/03/2022"
          },
          "endDate": {
            "type": "string",
            "example": "01/07/2022"
          }
        }
      },
      "SingleResult": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "example": 0.258
          }
        }
      }
    }
  }
}